add publish lanechange steering factor

+++ planning/behavior_path_planner/include/behavior_path_planner/planning_api_interface.hpp
@@ -37,7 +37,7 @@ public:
   PlanningAPIInterface(rclcpp::Node * node, const std::string & name);
   void publishSteeringFactor(const rclcpp::Time & stamp);
   void updateSteeringFactor(
-    const std::vector<Pose> & pose, const std::vector<float> distance, const uint16_t type,
+    const std::vector<Pose> & pose, const std::vector<double> distance, const uint16_t type,
     const uint16_t direction, const uint16_t status, const std::string detail);
   void clearSteeringFactors();
